+++ apps/rig/config/test.exs
@@ -2,3 +2,6 @@ use Mix.Config
 
 config :rig, Rig.Kafka,
   enabled?: true
+
+config :rig, Rig.KafkaTest,
+  topic: "rig"