+++ apps/rig/test/test_helper.exs
@@ -1,1 +1,3 @@
 ExUnit.start()
+# Exclude all smoke tests from running by default
+ExUnit.configure exclude: [smoke: true]