+++ apps/rig_auth/test/dummy_smoke_test.exs
@@ -1,0 +1,11 @@
+defmodule RigAuth.DummySmokeTest do
+  @moduledoc "Prevents a warning as long as there is no smoke test in this app."
+  use ExUnit.Case, async: true
+
+  describe "dummy" do
+    @tag :smoke
+    test "nothing" do
+      nil
+    end
+  end
+end