+++ apps/rig_inbound_gateway/lib/rig_inbound_gateway_web/endpoint.ex
@@ -16,11 +16,6 @@ defmodule RigInboundGatewayWeb.Endpoint do
 
   plug RigInboundGatewayWeb.Router
 
-  @doc """
-  Initialize the endpoint configuration.
-  Invoked when the endpoint supervisor starts, allows dynamically
-  configuring the endpoint from system environment or other runtime sources.
-  """
   @spec init(:supervisor, Keyword.t) :: {:ok, Keyword.t}
   def init(:supervisor, config) do
     Confex.Resolver.resolve(config)